MessagePatterns interface and XX pattern implementation
